+++ adminer/lang/fr.inc.php
@@ -286,7 +286,7 @@ Lang::$translations = array(
 	'Drop %s?' => 'Supprimer %s?',
 	'overwrite' => 'écraser',
 	'DB' => 'BD',
-	'ATTACH queries are not supported.' => 'Requêtes ATTACH ne sont pas supportées.',
+	'%s queries are not supported.' => 'Requêtes %s ne sont pas supportées.',
 	'Warnings' => 'Avertissements',
 	'Adminer does not support accessing a database without a password, <a href="https://www.adminer.org/en/password/"%s>more information</a>.' => 'Adminer ne supporte pas l\'accès aux bases de données sans mot de passe, <a href="https://www.adminer.org/en/password/"%s>plus d\'information</a>.',
 	'The action will be performed after successful login with the same credentials.' => 'Cette action sera exécutée après s\'être connecté avec les mêmes données de connexion.',